Added one dot, two dots stimulus
This stimulus is based on the stimulus described in (Gale & Murphy 2014).  dots move from top to bottom of screen. We either record one dot or two dots, at different locations. For two dots, 1 dot is always in the center (0 deg), the other is at a varying distance outside of the RF.